ci(stylelint): 修复string-quotes与prettier格式化冲突的问题

+++ stylelint.config.js
@@ -34,7 +34,7 @@ module.exports = {
     'no-descending-specificity': null,
 
     'function-url-quotes': 'always',
-    'string-quotes': 'double',
+    'string-quotes': 'single', // double跟prettier冲突
     indentation: 2,
     'unit-case': null,
     'color-hex-case': 'lower',
